+++ packages/cli/lib/featurePromptModules/typescript.js
@@ -1,6 +1,7 @@
 module.exports = (cli) => {
 	cli.injectFeature({
 		name: 'TypeScript',
+		checked: true,
 		value: 'ts',
 		short: 'TS',
 		description: 'Add support for the TypeScript language',