+++ packages/cli/lib/utils/setupDevProject.js
@@ -18,9 +18,11 @@ const linkBin = async (src, dest) => {
 		await fs.chmod(dest, '755') // 用于更改给定路径的权限
 	}
 }
+
 const setupDevProject = (targetDir, cliServiceName) => {
 	// 在node_modules/.bin 生成wa-cli-service 和wa-cli-service.cmd
 	// require.resolve("@wa-dev/cli-service/bin/wa-cli-service") 有软链
+	// require.resolve 会找node_modules里面的模块
 	return linkBin(
 		require.resolve(`@wa-dev/${cliServiceName}/bin/cli-service`),
 		path.join(targetDir, 'node_modules', '.bin', `wa-${cliServiceName}`)