+++ packages/cli-service/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @wa-dev/cli-service
 
+## 2.0.0-beat.9
+
+### Major Changes
+
+- 11a0260: fix: add eslint dependencies
+- 56756c9: test lost folder
+- 2650d71: fix template path error
+- 064a9a3: change devDependencies to dependencies
+- 3d3b2af: fix folder lost
+- 11a0260: fix some bugs
+- d3b1ca9: test fix miss webpack-dev-server
+
+### Patch Changes
+
+- Updated dependencies [11a0260]
+- Updated dependencies [56756c9]
+- Updated dependencies [2650d71]
+- Updated dependencies [064a9a3]
+- Updated dependencies [3d3b2af]
+- Updated dependencies [11a0260]
+- Updated dependencies [d3b1ca9]
+  - @wa-dev/cli-shared-utils@2.0.0-beat.9
+
 ## 2.0.0-beat.0
 
 ### Major Changes